requirements.txt
# $ conda create --name TopoTTA --file requirements.txt
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
bzip2=1.0.8=h5eee18b_6
ca-certificates=2025.9.9=h06a4308_0
certifi=2025.8.3=pypi_0
charset-normalizer=3.4.3=pypi_0
contourpy=1.3.3=pypi_0
cycler=0.12.1=pypi_0
expat=2.7.1=h6a678d5_0
filelock=3.19.1=pypi_0
fonttools=4.60.0=pypi_0
fsspec=2025.9.0=pypi_0
gudhi=3.11.0=pypi_0
hf-xet=1.1.10=pypi_0
huggingface-hub=0.35.3=pypi_0
idna=3.10=pypi_0
imageio=2.37.0=pypi_0
jinja2=3.1.6=pypi_0
joblib=1.5.2=pypi_0
kiwisolver=1.4.9=pypi_0
lazy-loader=0.4=pypi_0
ld_impl_linux-64=2.40=h12ee557_0
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libmpdec=4.0.0=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libuuid=1.41.5=h5eee18b_0
libxcb=1.17.0=h9b100fa_0
libzlib=1.3.1=hb25bd0a_0
markupsafe=3.0.2=pypi_0
matplotlib=3.10.6=pypi_0
mpmath=1.3.0=pypi_0
ncurses=6.5=h7934f7d_0
networkx=3.5=pypi_0
numpy=2.3.3=pypi_0
nvidia-cublas-cu12=12.8.4.1=pypi_0
nvidia-cuda-cupti-cu12=12.8.90=pypi_0
nvidia-cuda-nvrtc-cu12=12.8.93=pypi_0
nvidia-cuda-runtime-cu12=12.8.90=pypi_0
nvidia-cudnn-cu12=9.10.2.21=pypi_0
nvidia-cufft-cu12=11.3.3.83=pypi_0
nvidia-cufile-cu12=1.13.1.3=pypi_0
nvidia-curand-cu12=10.3.9.90=pypi_0
nvidia-cusolver-cu12=11.7.3.90=pypi_0
nvidia-cusparse-cu12=12.5.8.93=pypi_0
nvidia-cusparselt-cu12=0.7.1=pypi_0
nvidia-ml-py=13.580.82=pypi_0
nvidia-nccl-cu12=2.27.3=pypi_0
nvidia-nvjitlink-cu12=12.8.93=pypi_0
nvidia-nvtx-cu12=12.8.90=pypi_0
openssl=3.0.17=h5eee18b_0
packaging=25.0=pypi_0
pandas=2.3.2=pypi_0
pillow=11.3.0=pypi_0
pip=25.2=pyhc872135_0
pot=0.9.5=pypi_0
pthread-stubs=0.3=h0ce48e5_1
pynvml=13.0.1=pypi_0
pyparsing=3.2.4=pypi_0
python=3.13.7=h7e8bc2b_100_cp313
python-dateutil=2.9.0.post0=pypi_0
python_abi=3.13=1_cp313
pytz=2025.2=pypi_0
pyyaml=6.0.3=pypi_0
readline=8.3=hc2a1206_0
requests=2.32.5=pypi_0
safetensors=0.6.2=pypi_0
scikit-image=0.25.2=pypi_0
scikit-learn=1.7.2=pypi_0
scipy=1.16.2=pypi_0
setuptools=78.1.1=py313h06a4308_0
six=1.17.0=pypi_0
sqlite=3.50.2=hb25bd0a_1
sympy=1.14.0=pypi_0
threadpoolctl=3.6.0=pypi_0
tifffile=2025.9.9=pypi_0
timm=1.0.20=pypi_0
tk=8.6.15=h54e0aa7_0
torch=2.8.0=pypi_0
torchaudio=2.8.0=pypi_0
torchvision=0.23.0=pypi_0
tqdm=4.67.1=pypi_0
triton=3.4.0=pypi_0
typing-extensions=4.15.0=pypi_0
tzdata=2025.2=pypi_0
urllib3=2.5.0=pypi_0
wheel=0.45.1=py313h06a4308_0
xorg-libx11=1.8.12=h9b100fa_1
xorg-libxau=1.0.12=h9b100fa_0
xorg-libxdmcp=1.1.5=h9b100fa_0
xorg-xorgproto=2024.1=h5eee18b_1
xz=5.6.4=h5eee18b_1
zlib=1.3.1=hb25bd0a_0